main rendering loop done

+++ include/ppu.h
@@ -1,4 +1,4 @@
-void ppu_tick_to(int cycle);
+void ppu_tick_to(unsigned long long cycle);
 void ppu_setup(unsigned char *chrom, unsigned char *vram, unsigned char *palette, unsigned char *oam, int *nmi);
 unsigned char ppu_read(unsigned short addr);
 void ppu_write(unsigned short addr, unsigned char data);
